Move mime whitelist from GitlabManager to AssetGroup
The mime whitelist stuff doesn't seem to have anything to do with gitlab
so moving it instead to AssetGroup.
